requirements.txt
pyyaml>=5.1
requests>=2.21
